perf: 优化启动速度，弹窗和拼音库改为异步加载，删除启动占位文字
