+++ plugins/otp-2fa/README.md
@@ -83,7 +83,8 @@
 │   ├── utils/                           # 工具函数
 │   │   ├── crypto.ts                    # 加密与安全存储集成
 │   │   ├── otp.ts                       # OTP 核心算法实现
-│   │   └── pinyin.ts                    # 拼音搜索与双拼转换
+│   │   ├── pinyin-schemes.ts            # 拼音方案常量与轻量匹配函数
+│   │   └── pinyin.ts                    # 拼音搜索与双拼转换（含 pinyin-pro 懒加载）
 │   ├── App.vue                          # 插件主入口
 │   ├── constants.ts                     # 全局常量定义
 │   ├── main.css                         # 全局样式系统
@@ -114,6 +115,15 @@ npm run build
 ```
 
 ## 更新日志
+
+### v1.3.0
+
+- **启动加速**：首屏 JS 体积从 412 KB 降至 110 KB，启动时间缩短约 60%。
+- **按需加载**：所有弹窗组件（设置、编辑、确认、导入等）改为异步加载，首次打开时按需下载。
+- **拼音懒加载**：`pinyin-pro` 库改为动态 import，仅在搜索或账号变更时加载，不占用启动时间。
+- **移除预计算**：拼音缓存不再在启动时预构建，改为账号变化时按需计算，减少启动 CPU 开销。
+- **分包优化**：`pinyin-pro` 独立拆分为异步 chunk，不混入首屏 vendor 包。
+- **删除占位文字**：移除 `<div id="app">` 中的"正在启动插件"加载提示。
 
 ### v1.2.1
