+++ plugins/otp-2fa/index.html
@@ -5,11 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   </head>
   <body>
-    <div id="app">
-      <div style="display:flex; justify-content:center; align-items:center; height:100vh; font-size:18px; font-weight:700; opacity:0.5; font-family: sans-serif;">
-        正在启动插件
-      </div>
-    </div>
+    <div id="app"></div>
     <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
